+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Added
+- Added possibility for use `doctrine:schema:update` with the ENUM's filed type.
+
 ## [0.1.0] 2018-09-11
 ### Added
 - First release of this component.